+++ RELEASE_NOTES.md
@@ -1,6 +1,7 @@
 *:
 * Added function dataframe_to_netcdf() to encode NetCDF from an existing dataframe
 * Updated the Convention to include QC flags in it
+* Added filter on QC in netcdf_to_dataframe() and cat.py CLI command
 
 1.3.2 :
 * Fixed IEA_PVPS after updates of convention : Time -> time